+++ src/components/cards/crowdfunding-card/components/informations.js
@@ -31,17 +31,19 @@ var _typography = require("../../../../helpers/utils/typography");
 
 var _colorsConfig = _interopRequireDefault(require("../../../../constants/colors-config"));
 
+var _screenConfig = require("../../../../constants/screen-config");
+
 var COMPONENT_GUTTER = (0, _typography.pxToRem)(10);
 
 var StyledContainer = _styledComponents.default.div.withConfig({
   displayName: "informations__StyledContainer",
   componentId: "qg8b01-0"
-})(["margin-top:", ";margin-bottom:", ";display:flex;flex-wrap:wrap;padding:0 ", ";line-height:1;"], (0, _typography.pxToRem)(20), (0, _typography.pxToRem)(5), COMPONENT_GUTTER);
+})(["margin-top:", ";margin-bottom:", ";display:flex;flex-wrap:wrap;padding:0;line-height:1;@media (min-width:", "){margin-top:", ";padding:0 ", ";}"], (0, _typography.pxToRem)(15), (0, _typography.pxToRem)(5), (0, _typography.pxToRem)(_screenConfig.ScreenConfig.M.min), (0, _typography.pxToRem)(20), COMPONENT_GUTTER);
 
 var StyledInfo = _styledComponents.default.div.withConfig({
   displayName: "informations__StyledInfo",
   componentId: "qg8b01-1"
-})(["margin-top:", ";margin-right:", ";&:last-child{margin-right:0;}"], (0, _typography.pxToRem)(5), (0, _typography.pxToRem)(30));
+})(["margin-right:", ";&:last-child{margin-right:0;}@media (min-width:", "){margin-right:", ";}"], (0, _typography.pxToRem)(15), (0, _typography.pxToRem)(_screenConfig.ScreenConfig.M.min), (0, _typography.pxToRem)(30));
 
 var StyledInfoLoading = _styledComponents.default.span.withConfig({
   displayName: "informations__StyledInfoLoading",